refactor: :gem: the huge migration
