+++ src/components/home/DailyVerse.tsx
