+++ src/components/ui/button.tsx
@@ -1,7 +1,7 @@
 import { cva, type VariantProps } from 'class-variance-authority';
 import * as React from 'react';
 import { Pressable } from 'react-native';
-import { TextClassContext } from '~/components/ui/text';
+import { TextClassContext } from '~/src/components/ui/text';
 import { cn } from '~/lib/utils';
 
 const buttonVariants = cva(
