+++ src/components/ui/card.tsx
@@ -1,7 +1,7 @@
 import type { TextRef, ViewRef } from '@rn-primitives/types';
 import * as React from 'react';
 import { Text, TextProps, View, ViewProps } from 'react-native';
-import { TextClassContext } from '~/components/ui/text';
+import { TextClassContext } from '~/src/components/ui/text';
 import { cn } from '~/lib/utils';
 
 const Card = React.forwardRef<ViewRef, ViewProps>(
