+++ src/components/ui/popover.tsx
@@ -3,7 +3,7 @@ import * as React from 'react';
 import { Platform, StyleSheet } from 'react-native';
 import Animated, { FadeIn, FadeOut } from 'react-native-reanimated';
 import { cn } from '~/lib/utils';
-import { TextClassContext } from '~/components/ui/text';
+import { TextClassContext } from '~/src/components/ui/text';
 
 const Popover = PopoverPrimitive.Root;
