+++ src/components/ui/tooltip.tsx
@@ -2,7 +2,7 @@ import * as TooltipPrimitive from '@rn-primitives/tooltip';
 import * as React from 'react';
 import { Platform, StyleSheet } from 'react-native';
 import Animated, { FadeIn, FadeOut } from 'react-native-reanimated';
-import { TextClassContext } from '~/components/ui/text';
+import { TextClassContext } from '~/src/components/ui/text';
 import { cn } from '~/lib/utils';
 
 const Tooltip = TooltipPrimitive.Root;
